requirements.txt
# Django>=3.0,<3.1
django==2.2.12
python-decouple==3.3
django-cors-headers==3.2.1
gunicorn==20.0.4
psycopg2-binary==2.8.5
djangorestframework==3.11.0
Pillow==6.2.2
django-jet==1.0.8
djangorestframework-camel-case==1.1.2
django-mptt==0.11.0
# django-cities-light==3.6.0

django-querycount==0.7.0
factory-boy==2.12.0
